+++ ModelService/Macromanagement/Types/MCTS.cs
@@ -1,0 +1,16 @@
+﻿using ModelService.Types;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ModelService.Macromanagement.Types
+{
+    public class MCTS
+    {
+        public Node RootNode { get; set; }
+
+
+    }
+}
