+++ ModelService/ModelService.csproj
@@ -69,6 +69,8 @@
     <Compile Include="Macromanagement\Algorithm.cs" />
     <Compile Include="Macromanagement\Macromanagement.cs" />
     <Compile Include="Macromanagement\Policy.cs" />
+    <Compile Include="Macromanagement\Types\MCTS.cs" />
+    <Compile Include="Macromanagement\Types\POMDP.cs" />
     <Compile Include="Micromanagement\Algorithm.cs" />
     <Compile Include="Micromanagement\Micromanagement.cs" />
     <Compile Include="Micromanagement\Policy.cs" />
@@ -79,6 +81,7 @@
     <Compile Include="Services\RService.cs" />
     <Compile Include="Types\Army.cs" />
     <Compile Include="Types\Coordinate.cs" />
+    <Compile Include="Types\Node.cs" />
     <Compile Include="Types\Player.cs" />
     <Compile Include="Types\Unit.cs" />
   </ItemGroup>
@@ -86,8 +89,6 @@
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Macromanagement\Types\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
