+++ ModelService/Types/Node.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ModelService.Types
+{
+    public class Node
+    {
+        public List<Node> Children { get; set; }
+
+
+    }
+}
